20250210_project_template/backend/app/requirements.txt
fastapi[standard]
prometheus-fastapi-instrumentator
prometheus-client
redis[hiredis]
psycopg2-binary
minio
pika